Backend/requirements.txt
fastapi[all]
tiktoken
requests
bs4
youtube_transcript_api
langchain_community
langchain_openai
langchain_google_vertexai
langchain_core
langchain_qdrant
langgraph
langgraph-checkpoint-postgres
python-dotenv
IPython

beautifulsoup4
google_cloud_aiplatform
openai
pandas

psycopg
psycopg_binary
psycopg2-binary
psycopg-pool
boto3
botocore
cloudscraper
python-docx
